+++ src/spellbot/models/guild_member.py
@@ -2,21 +2,16 @@ from __future__ import annotations
 
 from datetime import UTC, datetime
 from functools import partial
-from typing import TYPE_CHECKING, TypedDict
+from typing import TYPE_CHECKING
 
 from sqlalchemy import BigInteger, Column, DateTime, ForeignKey
 
 from . import Base, now
 
 if TYPE_CHECKING:
+    from spellbot.data import GuildMemberData
+
     from . import Guild, User  # noqa: F401
-
-
-class GuildMemberDict(TypedDict):
-    created_at: datetime
-    updated_at: datetime
-    user_xid: int
-    guild_xid: int
 
 
 class GuildMember(Base):
@@ -56,10 +51,12 @@ class GuildMember(Base):
         doc="The external Discord ID of the guild",
     )
 
-    def to_dict(self) -> GuildMemberDict:
-        return {
-            "created_at": self.created_at,
-            "updated_at": self.updated_at,
-            "user_xid": self.user_xid,
-            "guild_xid": self.guild_xid,
-        }
+    def to_data(self) -> GuildMemberData:
+        from spellbot.data import GuildMemberData  # allow_inline
+
+        return GuildMemberData(
+            created_at=self.created_at,
+            updated_at=self.updated_at,
+            user_xid=self.user_xid,
+            guild_xid=self.guild_xid,
+        )