+++ tests/models/test_award.py
@@ -1,8 +1,11 @@
 from __future__ import annotations
 
+from dataclasses import asdict
 from typing import TYPE_CHECKING
 
 import pytest
+
+from spellbot.data import GuildAwardData, UserAwardData
 
 if TYPE_CHECKING:
     from tests.fixtures import Factories
@@ -11,11 +14,13 @@ pytestmark = pytest.mark.use_db
 
 
 class TestModelAward:
-    def test_guild_award(self, factories: Factories) -> None:
+    def test_guild_award_to_data(self, factories: Factories) -> None:
         guild = factories.guild.create()
         guild_award = factories.guild_award.create(guild=guild)
 
-        assert guild_award.to_dict() == {
+        award_data = guild_award.to_data()
+        assert isinstance(award_data, GuildAwardData)
+        assert asdict(award_data) == {
             "id": guild_award.id,
             "guild_xid": guild_award.guild_xid,
             "count": guild_award.count,
@@ -27,7 +32,7 @@ class TestModelAward:
             "verified_only": guild_award.verified_only,
         }
 
-    def test_user_award(self, factories: Factories) -> None:
+    def test_user_award_to_data(self, factories: Factories) -> None:
         guild = factories.guild.create()
         user = factories.user.create()
         guild_award = factories.guild_award.create(guild=guild)
@@ -37,7 +42,9 @@ class TestModelAward:
             guild_award_id=guild_award.id,
         )
 
-        assert user_award.to_dict() == {
+        award_data = user_award.to_data()
+        assert isinstance(award_data, UserAwardData)
+        assert asdict(award_data) == {
             "user_xid": user_award.user_xid,
             "guild_xid": user_award.guild_xid,
             "guild_award_id": user_award.guild_award_id,